requirements.txt
coloredlogs==15.0.1
faiss-cpu==1.8.0.post1
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
humanfriendly==10.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
onnx==1.16.1
onnxruntime==1.18.1
packaging==24.1
pillow==10.4.0
protobuf==5.27.2
sympy==1.13.0
torch==2.3.1
torchvision==0.18.1
triton==2.3.1
typing_extensions==4.12.2
